refactor(dsp): remove negotiation spi
