+++ data-protocols/dsp/dsp-negotiation/dsp-negotiation-transform/build.gradle.kts
@@ -18,7 +18,6 @@ plugins {
 
 dependencies {
     api(project(":data-protocols:dsp:dsp-http-spi"))
-    api(project(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-spi"))
     api(project(":data-protocols:dsp:dsp-transform"))
     api(project(":extensions:common:json-ld"))
     api(project(":spi:common:transform-spi"))
