+++ data-protocols/dsp/dsp-negotiation/dsp-negotiation-transform/src/main/java/org/eclipse/edc/protocol/dsp/negotiation/transform/DspNegotiationPropertyAndTypeNames.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.protocol.dsp.negotiation.spi;
+package org.eclipse.edc.protocol.dsp.negotiation.transform;
 
 public interface DspNegotiationPropertyAndTypeNames {
