+++ settings.gradle.kts
@@ -69,7 +69,6 @@ include(":data-protocols:dsp:dsp-catalog:dsp-catalog-transform")
 include(":data-protocols:dsp:dsp-transfer-process:dsp-transfer-process-transform")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-api")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-http-dispatcher")
-include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-spi")
 include(":data-protocols:dsp:dsp-negotiation:dsp-negotiation-transform")
 include(":data-protocols:dsp:dsp-http-core")
 include(":data-protocols:dsp:dsp-http-spi")
